requirements.txt
functions-framework==3.*
requests
pandas
haversine
openai
flask
python-dotenv